+++ CHANGES.md
@@ -11,11 +11,11 @@
 
 ## develop
 
-- [CHANGE] E2EE 機能を削除する
-  - @voluntas
+- [UPDATE] メッセージングの compress で fflate を利用せず decompressionStream API を利用するようにする
 
 ### misc
 
+- [CHANGE] E2EE 機能を削除する
 - [ADD] サンプルに getStats を追加する
   - sendrecv / sendonly / recvonly / messaging / simulcast
   - dataset に getStats の結果を入れて E2E テストで利用しやすくする